+++ CHANGELOG.md
@@ -18,6 +18,8 @@
     (GET "/noon-route" [] (ok "it's noon")))
 ```
 
+* Remove restructuring migration helpers for `1.0.0` (for `:parameters` and `:middlewares`)
+
 ## 1.2.0-alpha8 (18.5.2017)
 
 * **BREAKING**: `resource` function is always 1-arity, options and info are merged.